PCI: Change namespace from "pci::pci" to "pci".
